done-but-need-testing

+++ src/main/java/com/sila/model/OrderItem.java
@@ -19,7 +19,7 @@ public class OrderItem {
     @ManyToOne
     private Food food;
     private  int quantity;
-    private double totalPrice;
+    private Long totalPrice;
     private List<String> ingredients;
 
 }
